Hoist structured source-wiring into an OwnedStructuredJob hierarchy
StructuredResultsJob and StructuredReduceJob had a byte-identical
HandleSubscriptions override (owner.Sources = allSources.Distinct().ToArray()).
Rather than push a nullable owner plus an "is the owner set?" check onto the
shared base -- which StructuredRaceJob would inherit unused -- introduce an
intermediate OwnedStructuredJob<T> : StructuredJobBase<T> that owns the whole
source-wiring cluster: the non-null owner, allSources, Lock,
AccumulateSourcesAndObservers, and the HandleSubscriptions that publishes the
captured dependencies onto the owner.

ConcurrentMap's and ConcurrentMapReduce's jobs derive from it; StructuredRaceJob
stays on StructuredJobBase (it captures eagerly during evaluation and owns no
node) with the base's no-op HandleSubscriptions. This also drops the per-call
owner parameter (now a field) and the jobs' @this fields. Behavior-preserving.
Pinned by ConcurrentMap/MapReduce_TwoChildrenReadSameSignal_WiresItOnce plus the
ObservesSignalsReadByParallelChildren and ZeroFnStructuredNodes tests.
